Feature/gridedit 591 64 to 32 bit indexing 2 (#198)
Replaced size_t with UInt (32 bit unsigned ineger) (#198 | GRIDEDIT-591)

+++ libs/MeshKernel/include/MeshKernel/BilinearInterpolationOnGriddedSamples.hpp
@@ -44,8 +44,8 @@ namespace meshkernel
         /// @param[in] cellSize The grid cell size
         /// @param[in] values The values of the gridded samples
         BilinearInterpolationOnGriddedSamples(const Mesh2D& mesh,
-                                              size_t numXCoord,
-                                              size_t numYCoord,
+                                              UInt numXCoord,
+                                              UInt numYCoord,
                                               const Point& origin,
                                               double cellSize,
                                               const std::vector<double>& values);
@@ -81,18 +81,17 @@ namespace meshkernel
 
         /// @brief Gets the sample value at specific row and column
         /// @return The sample value
-        [[nodiscard]] double getGriddedValue(size_t columnIndex, size_t rowIndex) const
+        [[nodiscard]] double getGriddedValue(UInt columnIndex, UInt rowIndex) const
         {
             const auto index = rowIndex * m_numXCoord + columnIndex;
             return m_values[index];
         }
 
         const Mesh2D& m_mesh; ///< Pointer to the mesh
-
-        size_t m_numXCoord; ///< The number of x coordinates of the gridded data
-        size_t m_numYCoord; ///< The number of y coordinates of the gridded data
-        Point m_origin;     ///< The coordinate of the origin
-        double m_cellSize;  ///< The grid cell size
+        UInt m_numXCoord;     ///< The number of x coordinates of the gridded data
+        UInt m_numYCoord;     ///< The number of y coordinates of the gridded data
+        Point m_origin;       ///< The coordinate of the origin
+        double m_cellSize;    ///< The grid cell size
 
         std::vector<double> m_xCoordinates; ///< The x coordinates of the grid
         std::vector<double> m_yCoordinates; ///< The y coordinates of the grid
