+++ libs/MeshKernel/include/MeshKernel/MeshRefinement.hpp
@@ -145,7 +145,7 @@ namespace meshkernel
         /// @brief Computes refinement masks (compute_jarefine_poly)
         ///        Face nodes, edge and edge lengths are stored in local caches. See Mesh2D.FaceClosedPolygon method
         /// @param face The face index
-        void ComputeRefinementMasksFromSamples(size_t face);
+        void ComputeRefinementMasksFromSamples(UInt face);
 
         /// Computes the edge refinement mask (comp_jalink)
         void ComputeEdgesRefinementMask();
@@ -153,24 +153,24 @@ namespace meshkernel
         /// @brief Finds the hanging nodes in a face (find_hangingnodes)
         /// @param[in] face The current face index
         /// @returns The number of hanging edges on the face, the number of hanging nodes and the number of edges to refine
-        void FindHangingNodes(size_t face);
+        void FindHangingNodes(UInt face);
 
         /// @brief Get the number of hanging nodes
         /// @returns The number of hanging nodes
-        size_t CountHangingNodes() const;
+        UInt CountHangingNodes() const;
 
         /// @brief Get the number of hanging nodes
         /// @returns The number of hanging nodes
-        size_t CountHangingEdges() const;
+        UInt CountHangingEdges() const;
 
         /// @brief Get the number of hanging nodes
         /// @param[in] face The current face index
         /// @returns The number of hanging nodes
-        size_t CountEdgesToRefine(size_t face) const;
+        UInt CountEdgesToRefine(UInt face) const;
 
         /// Deletes isolated hanging nodes(remove_isolated_hanging_nodes)
         /// @returns Number of deleted isolated hanging nodes
-        [[nodiscard]] size_t DeleteIsolatedHangingnodes();
+        [[nodiscard]] UInt DeleteIsolatedHangingnodes();
 
         /// @brief Connect the hanging nodes with triangles (connect_hanging_nodes)
         void ConnectHangingNodes();
@@ -183,37 +183,37 @@ namespace meshkernel
 
         /// @brief The refinement operation by splitting the face (refine_cells)
         /// @param[in] numEdgesBeforeRefinement Number of edges before the refinement
-        void RefineFacesBySplittingEdges(size_t numEdgesBeforeRefinement);
+        void RefineFacesBySplittingEdges(UInt numEdgesBeforeRefinement);
 
         /// @brief Compute if an edge must be refined based on the face location type and the Courant criteria
         /// @param edge The index of the edge to be refined
         /// @param faceLocationType The face location type
         /// @returns If the edge should be refined
-        bool IsEdgeToBeRefinedBasedFaceLocationTypeAndCourantCriteria(size_t edge, FaceLocation faceLocationType) const;
+        bool IsEdgeToBeRefinedBasedFaceLocationTypeAndCourantCriteria(UInt edge, FaceLocation faceLocationType) const;
 
         /// @brief Compute if an edge must be refined based on the face location type
         /// @param edge The index of the edge to be refined
         /// @param depthValues The depth value
         /// @returns If the edge should be refined based on Courant criteria
-        bool IsRefineNeededBasedOnCourantCriteria(size_t edge, double depthValues) const;
+        bool IsRefineNeededBasedOnCourantCriteria(UInt edge, double depthValues) const;
 
         /// @brief Compute the face location type based on the depths values on the nodes
         void ComputeFaceLocationTypes();
 
         RTree m_samplesRTree; ///< The sample node RTree
 
-        std::vector<int> m_faceMask;        ///< Compute face without hanging nodes (1), refine face with hanging nodes (2), do not refine cell at all (0) or refine face outside polygon (-2)
-        std::vector<int> m_edgeMask;        ///< If 0, edge is not split
-        std::vector<int> m_nodeMask;        ///< The node mask used in the refinement process
-        std::vector<size_t> m_brotherEdges; ///< The index of the brother edge for each edge
+        std::vector<int> m_faceMask;      ///< Compute face without hanging nodes (1), refine face with hanging nodes (2), do not refine cell at all (0) or refine face outside polygon (-2)
+        std::vector<int> m_edgeMask;      ///< If 0, edge is not split
+        std::vector<int> m_nodeMask;      ///< The node mask used in the refinement process
+        std::vector<UInt> m_brotherEdges; ///< The index of the brother edge for each edge
 
         /// Local caches
         std::vector<bool> m_isHangingNodeCache;       ///< Cache for maintaining if node is hanging
         std::vector<bool> m_isHangingEdgeCache;       ///< Cache for maintaining if edge is hanging
         std::vector<Point> m_polygonNodesCache;       ///< Cache for maintaining polygon nodes
-        std::vector<size_t> m_localNodeIndicesCache;  ///< Cache for maintaining local node indices
-        std::vector<size_t> m_globalEdgeIndicesCache; ///< Cache for maintaining edge indices
-        std::vector<size_t> m_refineEdgeCache;        ///< Cache for the edges to be refined
+        std::vector<UInt> m_localNodeIndicesCache;    ///< Cache for maintaining local node indices
+        std::vector<UInt> m_globalEdgeIndicesCache;   ///< Cache for maintaining edge indices
+        std::vector<UInt> m_refineEdgeCache;          ///< Cache for the edges to be refined
         std::vector<FaceLocation> m_faceLocationType; ///< Cache for the face location types
 
         RefinementType m_refinementType = RefinementType::WaveCourant; ///< The type of refinement to use
